Working 1.19.4 implementation

+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/toolhandler/MagneticHandler.java
@@ -75,8 +75,8 @@ public class MagneticHandler implements RunnableHandler {
 
 
 	public List<Entity> getNearbyEntities(int range, Predicate<Entity> predicate) {
-		BlockPos pos1 = new BlockPos(rootVec.x + range, rootVec.y + range, rootVec.z + range);
-		BlockPos pos2 = new BlockPos(rootVec.x - range, rootVec.y - range, rootVec.z - range);
+		BlockPos pos1 = new BlockPos((int) (rootVec.x + range), (int) (rootVec.y + range), (int) (rootVec.z + range));
+		BlockPos pos2 = new BlockPos((int) (rootVec.x - range), (int) (rootVec.y - range), (int) (rootVec.z - range));
 		return rootEntity.getWorld().getOtherEntities(rootEntity, new Box(pos1, pos2), predicate);
 	}
